V1 of the TreeEventSlicer
First version of the TreeEventSlicer.
Allows one to slice the tree on samples, branching events or samples and
branching events cobined. Creates the breaks between 2 events in the
middle of the tree and a break EPS away from the last event on the list.
Additionally tries to keep the numbers of events in each group as equal
as possible, e.g. for 3 intervals on 8 events it will create intervals
with 3 3 and 2 events, and for 10 events it will create 3 3 4.